compress-video/requirements.txt
pika
boto3
ffmpeg-python
redis
python-dotenv
requests
loguru
prometheus_client
Flask
